Codes become a little bit cleaner

+++ server.js
@@ -1,18 +1,18 @@
 const express = require("express");
-const app = express();
 const https = require("https");
 const fs = require("fs");
-const db = require("./bd");
+const db = require("./db");
 const path = require("path");
 const request = require("request");
-const { getFakePrize } = require("./storage");
+const { getFakePrize } = require("./fakePrizesStorage");
 const { prizesCronJob } = require("./cronJobs");
 const createShareForm = require("./shareForm");
 
+const app = express();
 app.use(express.json());
 
+// start adding prizes to prize queue
 const prizeQueue = [];
-
 prizesCronJob(prizeQueue);
 
 app.get("/app", function(req, res) {
@@ -20,33 +20,20 @@ app.get("/app", function(req, res) {
 });
 
 app.post("/app/api/v1/user", async function(req, res) {
-  // -> get uid +
-  // if user exists:
-  // --> update subscribed and liked fields +
-  // --> return users attempts +
-  // else:
-  // --> create user +
-  // -> return user attempts +
   const { uid, liked, subscribed, location } = req.body;
   let user = await db.getUserByUid(uid);
   if (user) {
     const { attempts } = await db.updateUser({ uid, liked, subscribed }, user);
     res.send({ attempts: attempts });
+    res.end();
     return;
   }
   const attempts = await db.createUser({ uid, liked, subscribed, location });
   res.send({ attempts: attempts });
+  res.end();
 });
 
 app.get("/app/api/v1/getprize/:uid", async function(req, res) {
-  // -> decrement user attempts +
-  // if stack of real prizes is not empty:
-  // --> get prize from there (if user location allows it) +
-  // --> set user as winner +
-  // --> add user and prize to winners table +
-  // else:
-  // --> get fake prize +
-  // -> return prize +
   const { uid } = req.params;
   db.useAttempt(uid);
   const user = await db.getUserByUid(uid);
@@ -54,24 +41,25 @@ app.get("/app/api/v1/getprize/:uid", async function(req, res) {
   const { attempts } = user;
   if (prizeQueue.length == 0 || user.won) {
     res.send(Object.assign(fakePrize, { rest_attempts: attempts }));
+    res.end();
     return;
   }
   const prize = prizeQueue[0];
   let { locations } = prize;
   locations = locations.split(", ");
   if (locations.includes(user.location) || locations == []) {
-    const prize = prizeQueue.shift();
+    prize = prizeQueue.shift();
     db.setWon(uid);
     db.addWinner(user.id, prize.id);
     res.send(Object.assign(prize, { rest_attempts: attempts }));
+    res.end();
   }
-  res.send(Onject.assign(fakePrize, { rest_attempts: attempts }));
+  res.send(Object.assign(fakePrize, { rest_attempts: attempts }));
+  res.end();
 });
 
 app.post("/app/api/v1/getShareForm", async function(req, res) {
   const { title, text, img_url, upload_url } = req.body;
-  const canvas = await createShareForm(title, text, img_url);
-  const buffer = canvas.toBuffer();
   const filePath = path.join(
     __dirname,
     "public",
@@ -79,9 +67,11 @@ app.post("/app/api/v1/getShareForm", async function(req, res) {
     "images",
     `-${title.split("/").join("")}-.png`
   );
-  fs.writeFileSync(filePath, buffer);
-  // const formData = new FormData();
-  // formData.append("photo", fs.createReadStream(filePath));
+  if (!fs.existsSync(filePath)) {
+    const canvas = await createShareForm(title, text, img_url);
+    const buffer = canvas.toBuffer();
+    fs.writeFileSync(filePath, buffer);
+  }
   request.post(
     {
       url: upload_url,
@@ -94,8 +84,6 @@ app.post("/app/api/v1/getShareForm", async function(req, res) {
 });
 
 app.get("/app/api/v1/setReposted/:uid", async function(req, res) {
-  // -> set user reposted field to true
-  // -> return user attempts
   const { uid } = req.params;
   let { reposted, attempts } = await db.getUserByUid(uid);
   if (!reposted) {
@@ -105,8 +93,7 @@ app.get("/app/api/v1/setReposted/:uid", async function(req, res) {
 });
 
 app.use(express.static(path.join(__dirname)));
-
-//in prod change certificzte to real one
+//in production, change certificate to real one
 https
   .createServer(
     {
@@ -115,8 +102,8 @@ https
     },
     app
   )
-  .listen(443, function() {
+  .listen(5000, function() {
     console.log(
-      "Example app listening on port 3000! Go to https://localhost:443/"
+      "Example app listening on port 3000! Go to https://localhost:5000/"
     );
   });
